application-backend-server/requirements.txt
flask==3.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
pymysql==1.1.0
